+++ app/src/pages/Trends/TrendsPage.js
@@ -1,12 +1,15 @@
 import React from 'react'
 import TrendsForm from '../../components/Trends/TrendsForm'
 import TrendsExamples from '../../components/Trends/Examples/TrendsExamples'
+import TrendsExamplesItemTopic from '../../components/Trends/Examples/TrendsExamplesItemTopic'
 import './TrendsPage.css'
 
 const TrendsPage = () => (
   <div className='TrendsPage page'>
-    <h1 className='TrendsPage__title'>Explore any crypto trend</h1>
-    <TrendsForm />
+    <h1 className='TrendsPage__title'>
+      See how often a crypto term is mentioned in social media
+    </h1>
+    <TrendsExamplesItemTopic />
     <TrendsExamples />
   </div>
 )